+++ src/collections/assignments/collection.js
@@ -1,0 +1,6 @@
+/*
+ * pwix:roles/src/collections/assignments/collection.js
+ */
+
+export const Assignments = {
+};