+++ src/collections/assignments/index.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:roles/src/collections/assignments/index.js
+ */
+
+export { Assignments } from './collection.js';