+++ src/collections/assignments/server/index.js
@@ -1,0 +1,8 @@
+/*
+ * pwix:roles/src/collections/assignments/server/index.js
+ */
+
+import { Assignments } from '../index.js';
+
+import './deny.js';
+import './publish.js';