+++ src/server/js/collections.js
@@ -1,0 +1,6 @@
+/*
+ * pwix:roles/src/server/js/collections.js
+ */
+
+import '../../collections/aliases/server/index';
+import '../../collections/assignments/server/index';